feat(lfm2): consolidate text decoder into MLXLMCommon namespace (WS-C #2, issue #168)
Lifts the LFM 2 hybrid attention/short-conv layer stack into a public
`LFM2` namespace under `MLXLMCommon/Models/LFM2.swift`. Both the LLM
target (`MLXLLM/Models/LFM2.swift`) and the VLM target
(`MLXVLM/Models/LFM2VL.swift`) consume the shared types.

Shared:
- `LFM2.Configuration` (handles top-level + nested rope_parameters
  decoding)
- `LFM2.Attention` (uses applyRotaryPosition for batch-cache compat)
- `LFM2.ShortConv` (SSMStateCache-backed hybrid layer)
- `LFM2.MLP`
- `LFM2.DecoderLayer`
- `LFM2.ModelInner` (embed → N hybrid blocks → norm)

LLM-side: `LFM2Model: LLMModel` shrinks to a thin wrapper around
`LFM2.ModelInner` + tied lm_head (`embedTokens.asLinear(out)`).
Back-compat typealias `LFM2Configuration = LFM2.Configuration`.

VLM-side: `Language.LanguageModel` keeps its `LMOutput` wrap +
conv-weight transpose sanitize, now wrapping `LFM2.ModelInner`. The
private `Language` enum's duplicated layer classes (LFM2Attention,
LFM2ShortConv, LFM2MLP, LFM2DecoderLayer, LFM2ModelInner) are gone.
`LFM2VLConfiguration.TextConfiguration` is now a typealias to
`LFM2.Configuration`. Preemptive prefill-sync `eval()` barrier added
to `LFM2VL.prepare` mirroring the Gemma 3 / Gemma 4 VLM fix from
issue #169 — defends against the same class of bug since the
pre-consolidation private `LFM2ModelInner` may have masked it.

Net diff: +405 / -667 (~262-line reduction).

Verification:
- swift build clean
- swift test --configuration release: 203/204 (the 1 failure is the
  env-driven InferenceBenchmark scaffold — pre-existing).
- LLM smoke `mlx-community/LFM2-1.2B-4bit --method simple`:
  "My name is AI Assistant..." (254 tok/s)
- VLM smoke `mlx-community/LFM2-VL-1.6B-4bit --method vision`:
  "The image shows a golden retriever..." (292 tok/s, validation
  passes — "dog" in 136-token output)
